Add button component and move function into utility
